refactor(core): extract block types to block.rs
Move Block, BlockHash, BlockSpentOutputs, BlockTreeEntry, Coin and
TransactionSpentOutputs to dedicated block module.
